Solve Circular dependency problem
